Make store/fetch work in the API explorer.
There were two problems: the first was a wrong order of arguments
in the UPDATE statement. The second was that the API explorer
didn't understand the API description properly, due to  flask-swagger
using v2.0 of the API (https://swagger.io/docs/specification/2-0) rather
than the newer version 3. I copied the output of /api/spec into
https://editor.swagger.io/ and edited until it was happy.